moved test to test_import
